server/pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.9.2"
fastapi = "^0.115.0"
uvicorn = "^0.30.6"
python-firebase = "^1.2"
pydantic-settings = "^2.5.2"
openai = "^1.47.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"